make-kamikaze-2.1.2.sh
Because 2.1.2 is a special case in the release process (i.e. based on a previous 2.1.1 image instead of a clean Ubuntu or Debian base for kernel compatibility reasons) it gets its own build script.

It's simple and makes assumptions about being on an untouched 2.1.1 image to execute properly. However it executes the various changes needed to upgrade the 2.1.1 image to 2.1.2.